refactor(utils): rename file
